♻️ refactor: separate ProjectCard component
